misc 64-bit
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@5304 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ LoupeWindow.m
@@ -17,7 +17,7 @@
 
 @implementation LoupeWindow
 
-- (id)initWithContentRect:(NSRect)contentRect styleMask:(unsigned int)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
+- (id)initWithContentRect:(NSRect)contentRect styleMask:(NSUInteger)aStyle backing:(NSBackingStoreType)bufferingType defer:(BOOL)flag
 {	
     //Call NSWindow's version of this function, but pass in the all-important value of NSBorderlessWindowMask
     //for the styleMask so that the window doesn't have a title bar
